requirements.txt
# Core Django and API
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Video processing
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Machine Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.207  # YOLOv8
mmcv-full==1.7.1
mmaction2==1.1.0

# Data processing
pandas==2.1.3
numpy==1.24.3

# Utilities
python-decouple==3.8
boto3==1.34.0  # For AWS S3 storage
django-storages==1.14.2

# Development
python-dotenv==1.0.0 